refactor(game): change directory compositions
