+++ service/app/src/entities/game/model/index.ts
@@ -9,11 +9,6 @@ export type {
   PresignedUrlItem,
   PresignedUrlResponse,
 } from "./game"
-export * from './gameCreationActions';
-export * from './gameCreationReducer';
-export * from './gameCreationSelectors';
-export * from './gameCreationState';
-export * from './gameCreationUtils';
 export type {
   GameCreateQuestion,
   GameCreateRequest,
@@ -24,5 +19,5 @@ export type {
   MyGameQueryParams,
   PresignedUrlRequest,
 } from "./gameRequest"
-export * from './useGameCreation';
+export * from './state/create';
 export { filterGamesByQuery, findCursorPosition, sortGames } from "./utils"
