+++ service/app/src/entities/game/model/state/create/state.ts
