+++ service/app/src/entities/game/model/state/index.ts
@@ -1,0 +1,1 @@
+// barrel file
